+++ src/main/java/com/blog/dao/TagDao.java
@@ -3,7 +3,6 @@ package com.blog.dao;
 import com.blog.entity.Tag;
 import org.apache.ibatis.annotations.Mapper;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
 /**
@@ -14,48 +13,48 @@ import java.util.List;
 public interface TagDao {
     /**
      * 保存标签
-     * @param tag
-     * @return
+     * @param tag 标签
+     * @return 状态值
      */
     int saveTag(Tag tag);
 
     /**
      * 得到标签
-     * @param id
-     * @return
+     * @param id 标签id
+     * @return 标签
      */
-    Tag getTag(Long id);
+    Tag getTag(Integer id);
 
     /**
      * 通过名字获取标签
-     * @param name
-     * @return
+     * @param name 名字
+     * @return 标签
      */
     Tag getTagByName(String name);
 
     /**
      * 得到所有标签
-     * @return
+     * @return 标签列表
      */
     List<Tag> getAllTag();
 
     /**
      * 首页展示博客标签
-     * @return
+     * @return 标签列表
      */
     List<Tag> getBlogTag();
 
     /**
      * 更新标签
-     * @param tag
-     * @return
+     * @param tag 标签
+     * @return 状态值
      */
     int updateTag(Tag tag);
 
     /**
      * 删除标签
-     * @param id
-     * @return
+     * @param id 标签id
+     * @return 状态值
      */
-    int deleteTag(Long id);
+    int deleteTag(Integer id);
 }
