+++ src/main/java/com/blog/service/FriendLinkService.java
@@ -11,43 +11,42 @@ public interface FriendLinkService {
 
     /**
      * 查询所有友链
-     * @return
+     * @return 友链列表
      */
     List<FriendLink> listFriendLink();
 
     /**
      * 友链新增
-     * @param friendLink
-     * @return
+     * @param friendLink 友链
+     * @return 状态值
      */
     int saveFriendLink(FriendLink friendLink);
 
     /**
      * 根据id查询友链
-     * @param id
-     * @return
+     * @param id 友链id
+     * @return 友链
      */
-    FriendLink getFriendLink(Long id);
+    FriendLink getFriendLink(Integer id);
 
     /**
      * 根据网址查询友链
-     * @param blogAddress
-     * @return
+     * @param blogAddress 友链网址
+     * @return 友链
      */
     FriendLink getFriendLinkByBlogAddress(String blogAddress);
 
     /**
      * 编辑修改友链
-     * @param friendLink
-     * @return
+     * @param friendLink 友链
+     * @return 状态值
      */
     int updateFriendLink(FriendLink friendLink);
 
     /**
      * 删除友链
-     * @param id
+     * @param id 友链id
      */
-    void deleteFriendLink(Long id);
-
+    void deleteFriendLink(Integer id);
 
 }
